third_party/WebKit/LayoutTests/fast/css/image-orientation/image-orientation-expected.txt

Apply image-orientation property and check computed style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS img.style.cssText is "image-orientation: from-image;"
PASS getComputedStyle(img).imageOrientation is "from-image"
PASS img.style.cssText is "image-orientation: 0deg;"
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS img.style.cssText is ""
PASS getComputedStyle(img).imageOrientation is "0deg"
PASS successfullyParsed is true

TEST COMPLETE
